Faeture/add pending transfer requests

+++ app/controllers/transfer_requests_controller.rb
@@ -1,14 +1,15 @@
 # frozen_string_literal: true
 
 class TransferRequestsController < UsersController
-  before_action :set_transfer_request, only: %w[show destroy]
-  before_action :set_truckload, only: %w[create new destroy]
-  before_action :set_agent, only: %w[create new destroy]
-  before_action :set_driver, only: %w[create new destroy]
-  before_action :set_bank_account, only: %w[create new destroy]
+  before_action :set_transfer_request, only: %w[show destroy update]
+  before_action :set_truckload, only: %w[show create new destroy update]
+  before_action :set_agent, only: %w[show create new destroy update]
+  before_action :set_driver, only: %w[show create new destroy update]
+  before_action :set_bank_account, only: %w[show create new destroy update]
 
   def index
-    @q = TransferRequest.accessible_by(current_ability)
+    @q = TransferRequest.includes([bank_account: :person], :truckload, :user, :enterprise)
+                        .accessible_by(current_ability)
                         .page(params[:page])
                         .ransack(params[:q])
 
@@ -19,31 +20,70 @@ class TransferRequestsController < UsersController
     @transfer_request = TransferRequest.new
   end
 
-  def show; end
-
   def create
     @transfer_request = TransferRequest.new(params_transfer_request)
 
     if @transfer_request.save
-      redirect_to transfer_requests_path
+      redirect_to(transfer_requests_path)
       flash[:success] = 'Solicitação de transferência cadastrada com sucesso.'
     else
       render :new
     end
   end
 
+  def update
+    params_update_transfer_request = case params.require(:commit)
+                                     when 'Aprovar'
+                                       params_approve_transfer_request
+                                     when 'Rejeitar'
+                                       params_reject_transfer_request
+                                     end
+
+    if params_update_transfer_request == params_approve_transfer_request
+      notice = 'Solicitação de transferência aprovada com sucesso.'
+      @transfer_request.status_cd = 'approved'
+    elsif params_update_transfer_request == params_reject_transfer_request
+      notice = 'Solicitação de transferência rejeitada com sucesso.'
+      @transfer_request.status_cd = 'rejected'
+    end
+
+    @transfer_request.balance_value_truckload -= @transfer_request.value
+
+    return unless @transfer_request.update(params_update_transfer_request)
+
+    redirect_to(pending_transfer_requests_path)
+    flash[:success] = notice
+  end
+
   def destroy
-    if @transfer_request.destroy
-      redirect_to transfer_requests_path
-      flash[:success] = 'Solicitação de transferência excluída com sucesso.'
+    can_destroy_transfer_request = true if current_user.roles.kind_masters.present? ||
+                                           TransferRequest.find(params[:id]).user == current_user
+
+    if TransferRequest.find(params[:id]).status_cd != 'pending'
+      redirect_to(transfer_requests_path)
+      flash[:danger] = 'Você não pode deletar solicitação que não esteja pendente.'
+      return
+    end
+
+    if can_destroy_transfer_request
+      if @transfer_request.destroy
+        @transfer_request.truckload.transfer_requests.last.balance_value_truckload += @transfer_request.value
+        redirect_to(transfer_requests_path)
+        flash[:success] = 'Solicitação de transferência excluída com sucesso.'
+      else
+        render :index
+      end
     else
-      render :index
+      redirect_to(transfer_requests_path)
+      flash[:danger] = 'Você não pode deletar a solicitação de transferência de terceiros.'
     end
   end
 
   def truckload_information
     @truckload = Truckload.find(params[:id])
-    @truckload_value_driver = @truckload.value_driver.to_currency
+    @balance_value_truckload = @truckload.value_driver
+    @balance_value = @balance_value_truckload - @truckload.transfer_requests.sum(&:value)
+    @formatted_balance_value_truckload = @balance_value.to_currency
     @agent = @truckload&.agent&.person || []
     @driver = @truckload.driver.person
     @driver_bank_account = @driver.bank_accounts
@@ -54,12 +94,25 @@ class TransferRequestsController < UsersController
                           end
     information = {
                     truckload: @truckload,
-                    truckload_value_driver: @truckload_value_driver,
+                    balance_value_truckload: @balance_value,
+                    formatted_balance_value_truckload: @formatted_balance_value_truckload,
                     agent_bank_account: @agent_bank_account,
                     driver_bank_account: @driver_bank_account
                   }
 
     render json: information
+  end
+
+  def pending
+    @search_params = params[:q]
+    transfer_request = TransferRequest.includes([bank_account: :person], :truckload, :user, :enterprise)
+    @q = if current_user.roles.kind_masters.present?
+           transfer_request.where(status_cd: :pending).ransack(@search_params)
+         else
+           transfer_request.where(status_cd: :pending, enterprise: current_user.enterprise).ransack(@search_params)
+         end
+    @transfer_requests = @q.result.page(params[:page]).per(15)
+    @pending_transfer_request_count = @q.result.count
   end
 
   private
@@ -102,18 +155,32 @@ class TransferRequestsController < UsersController
   def set_bank_account
     @bank_accounts = if current_user.roles.kind_masters.present?
                        BankAccount.all
-                      else
-                        BankAccount.all
+                     else
+                       BankAccount.all
                      end
+  end
+
+  def params_approve_transfer_request
+    params.require(:transfer_request)
+          .permit(:voucher)
+          .with_defaults(updated_by_id: current_user.id)
+  end
+
+  def params_reject_transfer_request
+    params.require(:transfer_request)
+          .permit(:reject_reason)
+          .with_defaults(updated_by_id: current_user.id)
   end
 
   def params_transfer_request
     truckload = Truckload.find(params.require(:transfer_request)[:truckload_id])
     driver = Driver.find(truckload.driver_id) if params.require(:transfer_request)[:driver] == '1'
     agent = Agent.find(truckload.agent_id) if params.require(:transfer_request)[:agent] == '1'
+    balance_value_truckload = truckload.value_driver - truckload.transfer_requests.sum(&:value)
 
     params.require(:transfer_request)
           .permit(:value,
+                  :observation,
                   :type_cd,
                   :method_cd,
                   :truckload_id,
@@ -121,6 +188,7 @@ class TransferRequestsController < UsersController
           .with_defaults(enterprise: current_user.enterprise,
                          driver: driver,
                          agent: agent,
-                         user: current_user)
+                         user: current_user,
+                         balance_value_truckload: balance_value_truckload)
   end
 end
